シーン遷移の処理を一つのクラスにまとめた
