+++ Assets/Script/Sequence.cs.meta
@@ -1,5 +1,5 @@
 fileFormatVersion: 2
-guid: da27d1bdaf5d68947b32b82dc624742b
+guid: ff5799cfc0edae64f80b93710f855eff
 MonoImporter:
   externalObjects: {}
   serializedVersion: 2
