--- Assets/Script/TitleTransition.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: c49fb7c25a664b044a6892ba8d2f0d0a
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
